+++ src/ng-skos.js
@@ -9,4 +9,4 @@
  * handle SKOS data. See the [API reference](#api) for module documentation.
  */
 angular.module('ngSKOS',[])
-.constant('ngSKOS.version', '0.0.10');
+.constant('ngSKOS.version', '0.0.11');